--- services/home/src/components/JoinStudy/DetailDescription/index.tsx
@@ -1,5 +1,0 @@
-import { DetailDescriptionText } from '../StudyDescription/style';
-
-export function DetailDescription({ detailDescription }: { detailDescription: string }) {
-  return <DetailDescriptionText>{detailDescription}</DetailDescriptionText>;
-}
