--- services/home/src/components/JoinStudy/DetailDescription/style.ts
@@ -1,11 +1,0 @@
-import { theme } from "@depromeet/theme";
-import styled from "@emotion/styled";
-
-
-export const DetailDescriptionText = styled.span`
-  font-weight: 500;
-  font-size: 16px;
-  color: ${theme.colors.white};
-  padding-top: 19px;
-  display: flex;
-`;
