--- services/home/src/components/JoinStudy/StudyDescription/index.tsx
@@ -1,10 +1,0 @@
-import { DescriptionBox, DescriptionText, DescriptionWrapper } from './style';
-
-export function StudyDescription({ descriptionText }: { descriptionText: string }) {
-  return (
-    <DescriptionWrapper>
-      <DescriptionBox />
-      <DescriptionText>{descriptionText}</DescriptionText>
-    </DescriptionWrapper>
-  );
-}
